Fix the test failures on hppa
A couple of years ago the libffi port for hppa-linux was changed from using a
trampoline executed on the stack to the function descriptor technique used by
ia64. This doesn't require an executable stack and is simpler. However,
function pointers need to have the PLABEL bit set in the pointer. As a result,
a simple cast can't be used to convert closure pointers to function pointers.

python-cffi uses its own closure allocation so the problem can't be fixed in
libffi.

I added a macro CFFI_FN to do the conversion. It shouldn't affect other
architectures. There is a similar define in libffi.

Fixes: https://bugs.debian.org/1024271