🐛🧪 Fix referencing GHA artifact @ whl builders
It wasn't obvious right away, but referencing a job in the
`${{ needs }}` context that is not being depended on, and does not
even exist, silently returned an empty string for the whole expression
passed to the `name:` input of the `download-artifact` action
invocations. This subsequently triggered the behavior of said action to
change to "download all the artifacts that exist in the workflow".
This "download all" behavior has an additional quirk, though — it
downloads each artifact in a subdirectory named after the artifact name.
The behavior of the "download one specific artifact" mode, on the other
hand, is different in that it does not create an extra level of nesting.
This was the reason why it felt necessary to unpack two-levels deep
sdist tarballs in the first place.

With this patch, references to said sdists will change.