Unify common logic for dlopen(str)
With the clarity that the input filename has to be a str object at this
stage, we can now bring together the common logic that the win32 and
posix branches took.

Notice how the win32 implementation previously clearer the error if
PyArgs_ParseTuple failed, only to try again with the "et" format variant
(now "U" as per the previous commit). As noted in the previous comment
this was moot, both because the filename argument is guaranteed to be a
string at this point, and because it was used in PyUnicode_AsUTF8, which
would have failed if it wasn't given a str object.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>