Simplify argument parsing logic
Instead of having different flavours of PyArgs_ParseTuple we can instead
use a single one where all arguments are optional, then check the type
of the first argument (when present).

The main difference in this new implementation is that instead of
raising a TypeError if the argument is not a str object in the final
case, we simply assert it, making it clear that the higher-level API
that calls us ensures this is the case.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>